FCollada: Fix some warnings. Remove test code from default Linux build. Add 'test' build target.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7047 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/fcollada/src/FCollada/FColladaTest/FCTestExportImport/FCTEIExtra.cpp
@@ -79,7 +79,7 @@ namespace FCTestExportImport
 		FailIf(a2 == NULL);
 		FailIf(a1 == a2);
 		PassIf(IsEquivalent(a1->GetValue(), FC("Squirrel")));
-		PassIf(IsEquivalent(FUStringConversion::ToUInt32(a2->GetValue()), 1002));
+		PassIf(IsEquivalent(FUStringConversion::ToUInt32(a2->GetValue()), (uint32)1002));
 
 		// Identify the base node leaves
 		PassIf(baseNode->GetChildNodeCount() == 2);
